+++ project/stm32f412_bb_bt/bt_phybusif_h4.c
@@ -73,7 +73,7 @@ uint8_t hw_uart_bt_init(uint32_t baud_rate)
 ******************************************************************************/
 void uart_bt_send(uint8_t *buf,uint16_t len)
 {
-    uint8_t index;
+    uint16_t index;
     for(index = 0; index < len ; index++)
     {
         /* Wait until the last send is complete, then send the data */
